@OPNA2608 2024-07-05
https://github.com/NixOS/nixpkgs/blob/706eef542dec88cc0ed25b9075d3037564b2d164/pkgs/data/fonts/input-fonts/default.nix#L92-L95

Since OfBorg doesn't wanna do it: CC @jtojnar @romildo