@Aleksanaa 2024-07-05
> Since OfBorg doesn't wanna do it

That's strange, I didn't find anything wrong with your commit message

@ofborg build input-fonts